+++ HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 
 * Fixed resolving of C/C++ std libs by Eclipse IDE
   (`issue #220 <https://github.com/platformio/platformio/issues/220>`_)
+* Fixed "ConnectionError" when PlatformIO SF Storage is off-line
 
 2.0.2 (2015-05-27)
 ------------------